+++ common/src/main/java/net/bettercombat/logic/PlayerAttackHelper.java
@@ -7,6 +7,7 @@ import net.bettercombat.api.WeaponAttributes;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.ShieldItem;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.Arrays;
 
@@ -55,6 +56,9 @@ public class PlayerAttackHelper {
             if (attributes != null && attributes.attacks() != null) {
                 int handSpecificComboCount = ((isOffHand && comboCount > 0) ? (comboCount - 1) : (comboCount)) / 2;
                 var attackSelection = selectAttack(handSpecificComboCount, attributes, player, isOffHand);
+                if (attackSelection == null) {
+                    return null;
+                }
                 var attack = attackSelection.attack;
                 var combo = attackSelection.comboState;
                 return new AttackHand(attack, combo, isOffHand, attributes, itemStack);
@@ -64,6 +68,9 @@ public class PlayerAttackHelper {
             WeaponAttributes attributes = WeaponRegistry.getAttributes(itemStack);
             if (attributes != null && attributes.attacks() != null) {
                 var attackSelection = selectAttack(comboCount, attributes, player, false);
+                if (attackSelection == null) {
+                    return null;
+                }
                 var attack = attackSelection.attack;
                 var combo = attackSelection.comboState;
                 return new AttackHand(attack, combo, false, attributes, itemStack);
@@ -74,6 +81,7 @@ public class PlayerAttackHelper {
 
     private record AttackSelection(WeaponAttributes.Attack attack, ComboState comboState) { }
 
+    @Nullable
     private static AttackSelection selectAttack(int comboCount, WeaponAttributes attributes, PlayerEntity player, boolean isOffHandAttack) {
         var attacks = attributes.attacks();
         attacks = Arrays.stream(attacks)
@@ -85,6 +93,9 @@ public class PlayerAttackHelper {
                 .toArray(WeaponAttributes.Attack[]::new);
         if (comboCount < 0) {
             comboCount = 0;
+        }
+        if (attacks.length == 0) {
+            return null;
         }
         int index = comboCount % attacks.length;
         return new AttackSelection(attacks[index], new ComboState(index + 1, attacks.length));